moved the sequel db-wrapper into Database::Connection. close #4
