+++ lib/data-import.rb
@@ -9,7 +9,6 @@ require 'data-import/dependency_resolver'
 require 'data-import/runner'
 require 'data-import/execution_plan'
 require 'data-import/dsl'
-require 'data-import/adapters/sequel'
 require 'data-import/database'
 require 'data-import/definition'
 require 'data-import/importer'
